feat: Update DataLine::try_from
I realized that trying to find the `;` starting from the end of the line
is more efficient due to the fact that the max length of the temperature
value is 5 characters (-99.9), but the length of the city name could be
much more that that (39 characters in the case of my measurements.txt
file).

Also, based on that knowledge, we can use pattern matching to determine
the position of the `;` between the last 6 bytes of the line.